backend/Prediction-website/requirements.txt
fastapi==0.95.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.0
email-validator==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6

